refactor: move styles to css files
