+++ packages/panel/components/content-editor/content-editor.module.css
@@ -1,0 +1,9 @@
+.content {
+  margin: 0;
+  padding: 0;
+  list-style-type: none;
+}
+
+.element {
+  margin-bottom: 32px;
+}
