+++ packages/panel/components/content-editor/index.ts
@@ -1,0 +1,1 @@
+export { ContentEditor } from './content-editor';
