+++ packages/panel/components/element-editor/element-editor.module.css
@@ -1,5 +1,6 @@
 .props {
   position: relative;
+  max-width: 500px;
 }
 
 .remove {
